Add scripts.html to organize assets, cleaup hugo.toml
